Add initial GKE/Knative Terraform config.
Quick and dirty Terraform to get a GKE cluster with Knative installed.
This was mostly taken from the Knative installation guide:
https://github.com/knative/docs/blob/master/install/Knative-with-GKE.md#creating-a-kubernetes-cluster

DNS configuration has not been added yet. But you can follow the app
deployment guide and that works:
https://github.com/knative/docs/blob/master/install/getting-started-knative-app.md

There's no Knative Terraform module, hence why I've used a bash script
for now. In the future it might be wise to make this more Terraform
    and...less a bash script.