push: deploy application from source directory
Adds 'push' command. The command pushes two containers to the configured
container registry. The first is a container that has the local
directory's files. This container is then used as a source for a knative
build in conjunction with the buildpacks build template to create a
deployable container (the second container). Knative serving is then
used to deploy the second container.

The command requires two flags:
--container-registry: The container registry to push and pull the
containers from.

--service-account: The k8s service-account to access the container
registry.

Unlike the `cf` CLI, `kf push` does not follow the deployment process.
It will give some logs on uploading the first container, however it does
not (yet) follow the logs during the build and serve phase. Therefore a
developer will need to monitor the status of the app via:

```
kubectl get pods --watch
```

Example:

```
$ kv push --container-registry gcr.io/project-name --service-account build-bot
```

Change-Id: I08c8104a477d439eb9e135522f1eb6d1778f2be9