push: display build logs while pushing
Previously, when a developer would push an application via `kf push`,
they would have to find the build and then fetch the corresponding logs.
`kf push` would also exit after the source code was finished being
uploaded (and not when the app was deployed).

This CL makes `cf push` tail the logs after applying the configuration
and block until the application is deployed.

Change-Id: Ibc3eb967f73bc7c06e800faa45a2f568bdd35ba0