push: push docker images with --docker-image flag
This CL allows users to deploy an image that is already created and
available in a container registry. This skips the upload and build step.

Example usage:

```
kf push app-name --docker-image gcr.io/some-project/some-image --service-account some-account
```

This CL also adds the path flag to allow a developer to specify the
location of the source code. The functionality was already in place, but
the flag was not exposed. The path can be both relative or absolute.

```
kf push app-name --container-registry some-registry --service-account some-account --path ./workspace/app
```

Change-Id: I81f0598bb76d4aa2a5aed31b3d70992be5398830