push: make service account optional
A service account is only necessary if knative does not have access to
the container registry. This is not always necessary.

This CL also adds the service acccount name to the revision
configuration. It was previously only used in the build step. This
implied that if the registry was private and required the service
account to download the container, it would have not worked.

Change-Id: I96010d741f5b20901f39b80fde3116248cce67ca