push: adds --env/-e flag to set env var
If set, the values will set/replace the environment variables in the app
configuration. The values are formatted as `KEY=VALUE`. Anything that
does not adhere to that format is an error.

If an app already has environment variables and the flag is provided,
the environment variables are replaced. If it is not, then the
environment variables are left as is.

Example usage:
```
kf push app-name --container-registry gcr.io/some-registry --env KEY1=VAL1 --env KEY2=VAL2
```

Change-Id: I0812fc41bb70422c43acec3caf1c749385cc86a4