apps: shows if an app is deleting
The `apps` command will show `Deleting` for a reason if the app's
configuration is being deleted. It will get this information by
requesting both the services and configuration. The configuration is
used to gather if the app is being deleted while the service is used to
gather information about the app.

This CL does NOT include anything to prevent push from trying to modify
an app while it is being deleted.

Change-Id: Ia439f0aef8d3f2580aced6d5bf6e120a6b1dbef8