Adds doctor command to kf.
Doctor is built in a way that mimics the API of go's testing and
works with our assertion library.

Additional tests can be added by making exiting clients implement
Diagnosable then adding them to the list of tests in the entrypoint.

Clients can run the full battery of tests or choose specific cases
based on their high level grouping. For example, apps.

Change-Id: Ia220424fea02caccdce09d395e0538f3c51ef3f4