Fixed VCAP_SERVICES, fixes #1
Our VCAP_SERVICES generator was missing functionality that would
make services compatible with CF. Now we're most of the way there
except for tags.

Most of this should go away once we build a webhook that injects
VCAP_SERVICES into pods. That service will watch k8s resources
and keep an internal cache so the values can be generated quickly
leading to a different architecture.

Change-Id: If0340f0eab8a46218ce134fa12896727ec6a73ef