cmd/kf: use plugins from CRD (#29)
The CLI will first read the CRD to figure out which commands to
override. The CLI will still have the base commands (e.g., push, apps,
etc...). If the CLI doesn't find a CRD it can still operate correctly.
Any command that the CRD lists will be available (and not just ones that
override a base commands).

This CL also creates a "samples" dir that includes the buildpacks and
plugins. The only plugin at this moment is for GKE. Its likely this will
need a new home at some point.

Change-Id: I848c5385df9cacc0cbcf5c01515a789f1a5319cd