all: adds apps end to end tests and sample apps
part of #14

Adds end to end tests to cover the commands:
- push
- delete
- apps
- set-env
- unset-env
- env
- doctor

It adds the "envs" and "echo" sample Go apps. These apps are used by the
integration tests.

The integration tests are ran when the environment variable
GCP_PROJECT_ID is set while running the tests. If it is not set, they
will be skipped. They can also be skipped with the "--short" flag. The
tests run against a running knative cluster and assume GKE.