created client generator (#70)
This PR creates a client generator for kf that we can use to create clients for specific CF concepts. It follows the Kubernetes model for handling extensions and provides a way to run mutations and filters before writes and reads.

This filtering and automatic application means we can map multiple CF concepts to the same Kubernetes object model, apply defaults, and enforce inavariants on the created objects.

An example client is included to test the autogenerator and auto-generated code.