Added the ability to upsert values.
This adds an upsert ability to generated clients. It does this by
listing the values on the cluster to see if any match the one with the
given name then merging and updating if that's the case otherwise
creating.

We can't use clever apimachinery functions to check error responses
because some of our upstream clients still use vendoring so the types
won't match.

There's also a bug fix for filtering based on labels and fieldselectors.