buildpacks: removes upload-buildpacks command (#144)
This command has been replaced with a script
(./hack/upload-buildpacks-stack.sh). This is so the dependency on
github.com/buildpack/{lifecycle,pack} can be removed. These deps were
causing conflicts with upgrading knative.

This CL also adds a buildpack/stacks directory that allows a user to
alter the stack the buidpack uses.