pkg/kf/routes: adds routes commands
A route in KF is an Istio VirtualService. This CL adds commands that
manage the VirtualServices and names them routes. This CL does NOT bind
these routes to the apps (yet) and therefore does not complete the
routing story.

This CL adds the following commands:
- routes
- create-route
- delete-route

The remaining functionality will be added with the addition of the
`map-route` command.

related to #75