Feature/reconciler (#216)
Fixes #126

This adds a first pass at a controller. It follows Knative's lead with the way it builds controllers, logs, and reconciles.

The space definition is still in flux, but once it's complete we'll add unit-tests similar to Knative Serving's where each sets up one particular cluster status, then tests the controller's behavior.