pkg/kf/commands/apps: adds start and stop commands
The commands simply toggle the app.Spec.Instances.Stopped value. The
controller has been updated to delete the knative service if the app is
stopped. This is required because knative will otherwise delete the pods
and bring back a single pod for a while (even if scaling is set to 0):
https://github.com/knative/serving/issues/4098

fixes #282