Adds ability to poll until a predicate is done
This PR fixes #371 by borrowing concepts from the service catalog.
It adds WaitFor and WaitForE to every generated client which are
utilities to poll a k8s cluster until a predicate is true.

One predicate is provided for each client to allow the CLI to wait
for deletions.

Additionally, a function that checks if a duck type is completed
is added which resolves #377.