Remove options string retroactively. Fixes #211.
The template package we use does not let us adjust the internal formatting.
I was unable to find a way to disable options via the exposed settings.
However, we can capture cobra command output...and remove the troubled text.

In the future, we should move away from using this library altogether, especially
since it no longer exists in kubectl master