pkg/kf/commands/install: adds installer for GKE (#449)
The installer is an interactive installer that allows the user to setup
existing projects and clusters, or setup new ones. It then installs the
latest knative build and kf. Then finally it allows the user to create a
new space.

Note, it does not install a service broker (e.g., minibroker).

fixes #447