pkg/kf/commands/config: lazily return kube config errors
Before this CL, if a user does not have a $HOME/.kube/config, commands
like `kf version` and `kf install` would still fail.

This CL fixes that by returning a `rest.Config` that has a `Dial`
function that will always return an error if the config failed to load.

fixes #504