pkg/reconciler/route: recreate deleted VirtualServices (#537)
Prior to this CL, if a VirtualService was deleted, it was not properly
recreated. The reconciler would also panic in some situations when a
Route was not found.

fixes #521

This CL also fixes the case where the root path is registered. Before,
only '/' would match (and not '/index.html').

fixes #535