pkg/kf/commands/install: generate flags from installers
The installer will generate flags based on the installer. It will do a
single pass that won't run any commands and see what details the user
needs to provide. It will then generate flags from this and re-run the
installer. If a flag is set, the user won't be prompted to enter this
information.

It also adds a `--quiet -q` flag that instructs any yes/no questions to
be answered yes.

fixes #583