pkg/reconciler/app/resources: Set values to match ksvc defaults (#620)
Before this CL, the app reconciler would submit a Knative service and
the Knative webhook would alter the service. This would cause the kf
reconciler to kick back off and notice a delta. Therefore it would try
to update the app and fail with a conflict error.

It would eventually sort itself out, but not before confusing the kf
CLI.

Ideally, we could simply apply the defaults, however that results in an
error as well that we are mutating immutable fields.

fixes #618