pkg/reconciler/app: gc after revision is ready
Previously, the kf revision GC would remove all but the first generation
of knative revisions. This would imply that if a revision was slow to be
ready, then there would be downtime after something like a push.

This CL makes the GC ensure it is leaving a ready revision.

fixes #609