pkg/reconciler/space: don't reconcile ServiceAccounts secrets
Prior to this CL, the space reconciler would create a ServiceAccount
without a Secret. A different reconciler would then add a Secret, which
would cause our reconciler to remove it. This conflict caused the other
reconciler to create a new secret each iteration.

In general, our reconciler shouldn't care if a new secret is added to
the ServiceAccount. This CL removes all assertions on the Secrets. In
the future it should ensure the Kf secret is present, but not care if
there are extras.

fixes #809