pkg/apis/kf/v1alpha1: update service defaults (#810)
Previously, when pushing, the kf CLI would fail with the following
error:

```
[build] Updated state to: Error occurred while updating existing Knative
Service: Operation cannot be fulfilled on services.serving.knative.dev
"myapp": the object has been modified; please apply your changes to the
latest version and try again
```

A similar message was also reported by the controller.

This was caused by a mismatch in the desired Service Kf was creating and
what Knative Serving was defaulting. The mismatch would cause the
reconciler to try and update multiple times. This CL updates the
defaults.

fixes #772