pkg/kf/cli: adds interactive cli package
This CL adds the pkg/kf/cli package that has the interactive framework.
This framework builds cobra.Commands that allows a user to walk through
the command instead of configuring it only through flags.

The `kf install gke` command uses this framework to allow a user to
either configure a cluster through flags or through the existing
interactive UX.

fixes #547