pkg/kf/commands/install/gke: install configmap secrets
Prior to this CL, the installed cluster could not properly create a
space as it would be missing the required secret. This CL creates a
secret and configmap before asking the user to create the space.

It also sorts projects, clusters, machine types, zones and networks when
prompting the user.

fixes #885
fixes #836