cmd/controller-init: add config-secrets if not present (#896)
Previously, the controller would fail because it assumed the
config-secrets ConfigMap was available in the `kf` namespace. When a
user first installs kf, this won't be the case. Therefore the controller
enters a crash loop.

This CL fixes this by having the controller-init create the missing
ConfigMap if it's missing. It will not create the secret as it doesn't
have the capacity to properly fill it in. Therefore an operator will
still be require to fill it in properly (e.g.,
./hack/create-gke-secret.sh) before creating a space.

fixes #890