Added the ability to set labels and annotations on Apps for sticky sessions.
CF introduced the ability to set labels and annotations on apps with the
CLI v7, these match the metadata on K8s resources and are used for similar
purposes.

This PR adds the ability to set labels and annotations and updates the
controller to propagate them down to the application pods.

Additionally, we allow developers to create services to "break-glass"
which will allow advanced load balancing and networking techniques.

Change-Id: If9542a2f7eef09da36810d23b39122240f02ef84