Added CPU limits.
This change reverts the previous change that forced apps defining a CPU in the manifest to use that as both lower and upper bound.

It restores the behavior that by default apps will have no upper-bound and instead will only get the lower-bound if they set CPU.

Apps that want to set an upper-bound should use the new cpu-limit property instead.

There's also a new document outlining how resources work in Kf to
explain the caveats around setting CPUs.