Allow start commands to be displayed.
This fixes the existing start command behavior which panics.

It additionally adds the following:

* Tests for the different start command lookup scenarios.
* A way to enable/disable the behavior.

I've validated by installing in a cluster and trying with apps
with the global value enabled and disabled and was able to
reproduce the panic in the old code with the tests and confirm
that it no longer occurs.