Change the release builder to build ko.
Despite using an upgraded go image in certain steps, the ko
image that Kf was using to build still had an outdated version of
go (1.17) meaning that there was an inconsistency between local
development and what was being produced.

This change pivots us to use the latest version of ko on the
current go base image the pipeline uses ensuring Kf will always
be built with an expected version of Go, allowing us to stay up to
date on features and patches.