Fix feature flag reconciler triggers.
The Feature Flag reconciler previously was building too long of
a queue by enqueueing _all_ namespaces when the config-defaults
configmap changed.

While it appears there wasn't any harm due to the reconciler
ignoring the keys, it was still leading to concerningly high
queue depths and a huge amount of unnecessary work for the
controllers.