Update package.json (#1039)
Remove old "tech-doc-hugo" name which was triggering Snyk vulnerability
scans. Replaced it with currenty docsy contents.

This was from an older version of docsy:
https://github.com/google/docsy-example/issues/45

Looks like this name was taken over at one point:
https://security.snyk.io/vuln/SNYK-JS-TECHDOCHUGO-5870392. The scan is
triggering even though the name is just the same (even if we aren't and
never were using that package).