Remove references to cluster task (#1089)
Migrated the build process to use standard Tekton Task resources instead
of the deprecated
[ClusterTask](https://github.com/tektoncd/pipeline/issues/4476).

## Proposed Changes

* Remove any references to ClusterTask, this involved:
* Updating build validation tests to use NamespacedTaskKind (Task)
instead of ClusterTask.
* Removing the obsolete logic for handling ClusterTask from the build
reconciler.
* Correcting outdated code comments that referred to ClusterTask.

## Release Notes

<!-- kf follows the Keep A Changelog standard for release notes.

https://keepachangelog.com/en/1.0.0/

Changelog entries should be one per line and start with one of the
following
words:

`Added` for new features.
`Changed` for changes in existing functionality.
`Deprecated` for soon-to-be removed features.
`Removed` for now removed features.
`Fixed` for any bug fixes.
`Security` in case of vulnerabilities.

If one of the changes is breaking include that as a second word e.g.

  Removed BREAKING support for manifests v1
-->

```release-note
Changed: Migrated the build process to use standard Tekton Task resources instead of the deprecated ClusterTask.
```